Script to remake the autoqaqc thresholds for the 2025 sensor data
Here is the script that I used to make the new sensor thresholds using the 2023 manually verified and post verified data, as well as "good-ish" 2024 data. Since the data folder re-org and the naming convention change I got a little lost in the sauce with where I should find this data. Please check that the data sources that I am pulling make sense. Feel free to run the script, though you will overwrite the manual thresholds (nothing will be lost). If there are more sites or parameters that we want to add to this threshold list it should be very easy to add them via the site and parameter levels objects and perhaps some tweaks here and there. The output is in `"data/field_notes/qaqc/updated_seasonal_thresholds_2025.csv"`.